requirements.txt
playwright>=1.40.0
pytest>=7.4.0
pytest-html>=4.1.0
pytest-xdist>=3.5.0
pytest-timeout>=2.2.0
pytest-asyncio>=0.23.0
pyyaml>=6.0.1
requests>=2.31.0
httpx>=0.25.0
allure-pytest>=2.13.2
jsonschema>=4.20.0
faker>=20.0.0
python-dotenv>=1.0.0
